SENA-090 Viewing Total Sales by Day and Month with Profit Calculation - Frontend

+++ LaCiguenia-Backend/src/main/java/LaCiguenia/component/invoice/IInvoiceComponent.java
@@ -4,6 +4,6 @@ import LaCiguenia.commons.domains.entity.invoice.InvoiceEntity;
 import java.util.List;
 
 public interface IInvoiceComponent {
-    Integer invoiceSalesTotalDay (List<InvoiceEntity> listInvoice);
-    Integer invoiceSalesTotalMonth (List<InvoiceEntity> listInvoice);
+    Double invoiceSalesTotalDay (List<InvoiceEntity> listInvoice);
+    Double invoiceSalesTotalMonth (List<InvoiceEntity> listInvoice);
 }
